+++ packages/solid/src/segment-group/index.ts
@@ -3,16 +3,16 @@ import { useSegmentContext } from './segment-context'
 import { SegmentControl, type SegmentControlProps } from './segment-control'
 import { SegmentGroup as SegmentGroupRoot, type SegmentGroupProps } from './segment-group'
 import { useSegmentGroupContext } from './segment-group-context'
+import { SegmentGroupIndicator, type SegmentGroupIndicatorProps } from './segment-group-indicator'
 import { SegmentGroupLabel, type SegmentGroupLabelProps } from './segment-group-label'
 import { segmentGroupAnatomy } from './segment-group.anatomy'
-import { SegmentIndicator, type SegmentIndicatorProps } from './segment-indicator'
 import { SegmentLabel, type SegmentLabelProps } from './segment-label'
 
 const SegmentGroup = Object.assign(SegmentGroupRoot, {
   Root: SegmentGroupRoot,
   Label: SegmentGroupLabel,
+  Indicator: SegmentGroupIndicator,
   Segment: Segment,
-  SegmentIndicator: SegmentIndicator,
   SegmentLabel: SegmentLabel,
   SegmentControl: SegmentControl,
 })
@@ -21,8 +21,8 @@ export {
   Segment,
   SegmentControl,
   SegmentGroup,
+  SegmentGroupIndicator,
   SegmentGroupLabel,
-  SegmentIndicator,
   SegmentLabel,
   segmentGroupAnatomy,
   useSegmentContext,
@@ -31,9 +31,9 @@ export {
 
 export type {
   SegmentControlProps,
+  SegmentGroupIndicatorProps,
   SegmentGroupLabelProps,
   SegmentGroupProps,
-  SegmentIndicatorProps,
   SegmentLabelProps,
   SegmentProps,
 }